Add refactor to reduce the size of the city information facade
