+++ app/controllers/locations_controller.rb
@@ -1,6 +1,7 @@
 # frozen_string_literal: true
 
 class LocationsController < ApplicationController
+  helper_method :city
   
   def show
     city
