+++ app/views/shared/_navbar.html.erb
@@ -10,9 +10,9 @@
         <li class="nav-item active">
           <%= link_to 'Home', root_path, class: 'nav-link' %>
         </li>
-        <li class="nav-item active">
-          <%= link_to 'Locations', search_path, class: 'nav-link' %>
-        </li>
+        <%# <li class="nav-item active"> %>
+          <%# <%= link_to 'Locations', search_path, class: 'nav-link' %> 
+        <%# </li> %>
         <% if current_user%>
         <li class="nav-item active">
           <%= link_to 'Dashboard', dashboard_path, class: 'nav-link' %>
