+++ app/views/users/show.html.erb
@@ -1,2 +1,23 @@
 
-<h5>Welcome, <%= current_user.name %></h5>
+<nav>
+  <ol class="breadcrumb">
+    <li class="breadcrumb-item">
+      <%= link_to 'Home', root_path %>
+    </li>
+    <li class="breadcrumb-item active">Dashboard</li>
+  </ol>
+</nav>
+<h1>Welcome <%= @user.to_s %></h1>
+<h4>Active Favorites: <%= current_user.favorites.count %></h4>
+
+<div class="list-group">
+  <% current_user.favorites.all.each do |favorite| %>
+    <%= link_to "/city?q=#{favorite.slug}&id=#{favorite.geo_id}",
+                class: 'list-group-item list-group-item-action favorites' do %>
+      <div class="d-flex w-100 justify-content-between">          
+        <h5 class="mb-1"><%= favorite.name %></h5>
+        <small>Saved <%= time_ago_in_words(favorite.updated_at) %> ago</small>
+      </div>
+    <% end %>
+  <% end %>
+</div>
