+++ db/seeds.rb
@@ -5,3 +5,7 @@
 #
 #   movies = Movie.create([{ name: 'Star Wars' }, { name: 'Lord of the Rings' }])
 #   Character.create(name: 'Luke', movie: movies.first)
+
+user = User.create(name: 'FunyonFreak96', email: 'me@you.com', google_token: 'abc123', uid: '12')
+location = Location.create(geo_id: '16000US0820000')
+favorite = user.locations << location
